Creo la clase que maneja los tiles interactivos como hazards

+++ core/src/com/mygdx/vault/Vault.java
@@ -8,8 +8,8 @@ import com.badlogic.gdx.utils.ScreenUtils;
 import com.mygdx.vault.Screens.PlayScreen;
 
 public class Vault extends Game {
-	public static final int V_WIDTH = 17408;
-	public static final int V_HEIGHT = 9216;
+	public static final float V_WIDTH = 17408;
+	public static final float V_HEIGHT = 9216;
 	public static  final float PPM = 200;//pixeles que se mueve un objeto por metro en box2d
 
 	//SpriteBatch contiene todos los sprites, que luego se muestran
